+++ docker-image/src/FAISS_retriever.py
